gcloud/rainbow/requirements.txt
gunicorn
flask
flask_restful
google-cloud-firestore
wtforms